requirements.txt
torch>=2.0
transformers>=4.36
diffusers>=0.25
accelerate>=0.28
datasets>=2.19
Pillow>=10.0
tqdm>=4.66
